pyproject.toml
[tool.poetry]
name = "simple"
version = "0.1.0"
description = "Simple Hello World project."
authors = ["OpenAI <support@openai.com>"]
readme = "README.md"
packages = [{ include = "simple" }]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^2.7.1"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
simple = "simple.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
